Layering on top of the base image causes a version mismatch with CI.
 - Everytime we build, we need to create a new version,
   for the baseimage,and this new version must be present
   in /usr/lib/os-release as part of the OS, and
   right LABEL added as well.
-  Create a new version based in the baseimage
   version so we can track from each baseimage it
   was layered from.
- Also add a numeric suffix (as done for the FCOS
  stream) to identify the stream from the version:
    - 01: NVIDIA stream
    - 01: first build
  Example: 9.6.20260217-0101

- TODO: We should probably merged it with the
f-c-c versionary file, so we don't need to keep
our own version

+++ Containerfile
@@ -1,8 +1,21 @@
-# This builds the OCP4NV node image on top of the base image itself.
+# This builds the OCP4NV nodeimage/baseimage on top of the nodeimage/baseimage itself.
 ARG FROM
+ARG VERSION
+ARG TYPE
 FROM ${FROM}
-COPY voyager.repo /etc/yum.repos.d/
+
+ARG VERSION
+ARG TYPE
+COPY voyager.repo  /etc/yum.repos.d/
+
 RUN <<EOF
+
+if [ "$TYPE" = "baseimage" ]; then
+    OLD_VERSION=$(grep "^OSTREE_VERSION=" /usr/lib/os-release | cut -d"'" -f2)
+    [ -n "$OLD_VERSION" ] || { echo "Error: Could not extract version from os-release"; exit 1; }
+    sed -i "s|$OLD_VERSION|$VERSION|g" /usr/lib/os-release
+fi
+
 arch="$(uname -m)"
 dnf remove kernel-core --assumeyes
 
@@ -14,3 +27,5 @@ dnf install --assumeyes --disablerepo=* --enablerepo=voyager "$kernel_name"
 kver=$(ls /usr/lib/modules)
 env DRACUT_NO_XATTR=1 dracut -vf /usr/lib/modules/$kver/initramfs.img "$kver"
 EOF
+
+LABEL org.opencontainers.image.version=$VERSION
